add backstop gauge, backstop edits
