+++ src/pages/api/vebao/claimable-fees.ts
@@ -1,0 +1,32 @@
+import { useBlockUpdater } from '@/hooks/base/useBlock'
+import { useTxReceiptUpdater } from '@/hooks/base/useTransactionProvider'
+import { FeeDistributor } from '@/typechain/FeeDistributor'
+import { providerKey } from '@/utils/index'
+import { useQuery } from '@tanstack/react-query'
+import { useWeb3React } from '@web3-react/core'
+import useContract from '@/hooks/base/useContract'
+
+export const useClaimableFees = () => {
+	const { library, account, chainId } = useWeb3React()
+	const feeDistributor = useContract<FeeDistributor>('FeeDistributor')
+	const enabled = !!feeDistributor && !!account
+
+	const { data: claimableFees, refetch } = useQuery(
+		['@/hooks/vebao/useClaimableFees', providerKey(library, account, chainId), { enabled }],
+		async () => {
+			const _claimableFees = await feeDistributor.callStatic['claim(address)'](account)
+			return _claimableFees
+		},
+		{
+			enabled,
+		},
+	)
+
+	const _refetch = () => {
+		if (enabled) refetch()
+	}
+	useTxReceiptUpdater(_refetch)
+	useBlockUpdater(_refetch, 10)
+
+	return claimableFees
+}
